+++ crates/matrix-sdk-ui/src/timeline/pagination.rs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::ops::ControlFlow;
-
 use async_rx::StreamExt as _;
 use async_stream::stream;
 use futures_core::Stream;
@@ -21,9 +19,9 @@ use futures_util::{pin_mut, StreamExt as _};
 use matrix_sdk::event_cache::{
     self,
     paginator::{PaginatorError, PaginatorState},
-    BackPaginationOutcome, EventCacheError, RoomPagination,
+    EventCacheError, RoomPagination,
 };
-use tracing::{instrument, trace, warn};
+use tracing::{instrument, warn};
 
 use super::Error;
 
@@ -75,38 +73,30 @@ impl super::Timeline {
     ///
     /// Returns whether we hit the start of the timeline.
     async fn live_paginate_backwards(&self, batch_size: u16) -> event_cache::Result<bool> {
-        let pagination = self.event_cache.pagination();
-
-        let result = pagination
-            .run_backwards(
-                batch_size,
-                |BackPaginationOutcome { events, reached_start },
-                 _timeline_has_been_reset| async move {
-                    let num_events = events.len();
-                    trace!("Back-pagination succeeded with {num_events} events");
-
-                    if num_events == 0 && !reached_start {
-                        // As an exceptional contract: if there were no events in the response,
-                        // and we've not hit the start of the timeline, retry until we get
-                        // some events or reach the start of the timeline.
-                        return ControlFlow::Continue(());
+        loop {
+            match self.event_cache.pagination().run_backwards_once(batch_size).await {
+                Ok(outcome) => {
+                    // As an exceptional contract, restart the back-pagination if we received an
+                    // empty chunk.
+                    if outcome.reached_start || !outcome.events.is_empty() {
+                        return Ok(outcome.reached_start);
                     }
+                }
 
-                    ControlFlow::Break(reached_start)
-                },
-            )
-            .await;
+                Err(EventCacheError::BackpaginationError(
+                    PaginatorError::InvalidPreviousState {
+                        actual: PaginatorState::Paginating, ..
+                    },
+                )) => {
+                    // Treat an already running pagination exceptionally, returning false so that
+                    // the caller retries later.
+                    warn!("Another pagination request is already happening, returning early");
+                    return Ok(false);
+                }
 
-        match result {
-            Err(EventCacheError::BackpaginationError(PaginatorError::InvalidPreviousState {
-                actual: PaginatorState::Paginating,
-                ..
-            })) => {
-                warn!("Another pagination request is already happening, returning early");
-                Ok(false)
+                // Propagate other errors as such.
+                Err(err) => return Err(err),
             }
-
-            result => result,
         }
     }