+++ crates/matrix-sdk/src/event_cache/pagination.rs
@@ -14,7 +14,7 @@
 
 //! A sub-object for running pagination tasks on a given room.
 
-use std::{future::Future, ops::ControlFlow, sync::Arc, time::Duration};
+use std::{sync::Arc, time::Duration};
 
 use eyeball::Subscriber;
 use matrix_sdk_base::timeout::timeout;
@@ -45,76 +45,45 @@ impl RoomPagination {
     /// This automatically takes care of waiting for a pagination token from
     /// sync, if we haven't done that before.
     ///
-    /// The `until` argument is an async closure that returns a [`ControlFlow`]
-    /// to decide whether a new pagination must be run or not. It's helpful when
-    /// the server replies with e.g. a certain set of events, but we would like
-    /// more, or the event we are looking for isn't part of this set: in this
-    /// case, `until` returns [`ControlFlow::Continue`], otherwise it returns
-    /// [`ControlFlow::Break`]. `until` receives [`BackPaginationOutcome`] as
-    /// its sole argument.
-    ///
-    /// # Errors
-    ///
-    /// It may return an error if the pagination token used during
-    /// back-pagination has disappeared while we started the pagination. In
-    /// that case, it's desirable to call the method again.
-    ///
-    /// # Example
-    ///
-    /// To do a single run:
-    ///
-    /// ```rust
-    /// use std::ops::ControlFlow;
-    ///
-    /// use matrix_sdk::event_cache::{
-    ///     BackPaginationOutcome,
-    ///     RoomPagination,
-    ///     TimelineHasBeenResetWhilePaginating
-    /// };
-    ///
-    /// # async fn foo(room_pagination: RoomPagination) {
-    /// let result = room_pagination.run_backwards(
-    ///     42,
-    ///     |BackPaginationOutcome { events, reached_start },
-    ///      _timeline_has_been_reset: TimelineHasBeenResetWhilePaginating| async move {
-    ///         // Do something with `events` and `reached_start` maybe?
-    ///         let _ = events;
-    ///         let _ = reached_start;
-    ///
-    ///         ControlFlow::Break(())
-    ///     }
-    /// ).await;
-    /// # }
-    #[instrument(skip(self, until))]
-    pub async fn run_backwards<Until, Break, UntilFuture>(
+    /// It will run multiple back-paginations until one of these two conditions
+    /// is met:
+    /// - either we've reached the start of the timeline,
+    /// - or we've obtained enough events to fulfill the requested number of
+    ///   events.
+    #[instrument(skip(self))]
+    pub async fn run_backwards_until(
         &self,
-        batch_size: u16,
-        mut until: Until,
-    ) -> Result<Break>
-    where
-        Until: FnMut(BackPaginationOutcome, TimelineHasBeenResetWhilePaginating) -> UntilFuture,
-        UntilFuture: Future<Output = ControlFlow<Break, ()>>,
-    {
-        let mut timeline_has_been_reset = TimelineHasBeenResetWhilePaginating::No;
+        num_requested_events: u16,
+    ) -> Result<BackPaginationOutcome> {
+        let mut events = Vec::new();
 
         loop {
-            if let Some(outcome) = self.run_backwards_impl(batch_size).await? {
-                match until(outcome, timeline_has_been_reset).await {
-                    ControlFlow::Continue(()) => {
-                        trace!("back-pagination continues");
-
-                        timeline_has_been_reset = TimelineHasBeenResetWhilePaginating::No;
-
-                        continue;
-                    }
-
-                    ControlFlow::Break(value) => return Ok(value),
+            if let Some(outcome) = self.run_backwards_impl(num_requested_events).await? {
+                events.extend(outcome.events);
+                if outcome.reached_start || events.len() >= num_requested_events as usize {
+                    return Ok(BackPaginationOutcome {
+                        reached_start: outcome.reached_start,
+                        events,
+                    });
                 }
+                trace!("restarting back-pagination, because we haven't reached the start or obtained enough events yet");
             }
 
-            timeline_has_been_reset = TimelineHasBeenResetWhilePaginating::Yes;
+            debug!("restarting back-pagination because of a timeline reset.");
+        }
+    }
 
-            debug!("back-pagination has been internally restarted because of a timeline reset.");
+    /// Run a single back-pagination for the requested number of events.
+    ///
+    /// This automatically takes care of waiting for a pagination token from
+    /// sync, if we haven't done that before.
+    #[instrument(skip(self))]
+    pub async fn run_backwards_once(&self, batch_size: u16) -> Result<BackPaginationOutcome> {
+        loop {
+            if let Some(outcome) = self.run_backwards_impl(batch_size).await? {
+                return Ok(outcome);
+            }
+            debug!("restarting back-pagination because of a timeline reset.");
         }
     }
 
@@ -209,18 +178,16 @@ impl RoomPagination {
         // During a backwards pagination, when a duplicated event is found, the old
         // event is kept and the new event is ignored. This is the opposite strategy
         // than during a sync where the old event is removed and the new event is added.
-        //
-        // Let's forget the new events that are duplicated.
         if !all_deduplicated {
+            // Let's forget the new events that are duplicated.
             events.retain(|new_event| {
                 new_event
                     .event_id()
                     .map(|event_id| !duplicated_event_ids.contains(&event_id))
                     .unwrap_or(false)
             });
-        }
-        // All new events are duplicated, they can all be ignored.
-        else {
+        } else {
+            // All new events are duplicated, they can all be ignored.
             events.clear();
         }
 
@@ -416,16 +383,6 @@ impl From<Option<String>> for PaginationToken {
             None => Self::None,
         }
     }
-}
-
-/// A type representing whether the timeline has been reset.
-#[derive(Debug)]
-pub enum TimelineHasBeenResetWhilePaginating {
-    /// The timeline has been reset.
-    Yes,
-
-    /// The timeline has not been reset.
-    No,
 }
 
 #[cfg(test)]